feat: update usePcbQueue hook to unified useQueue hook
