+++ src/Pages/Rfq/RfqPage.tsx
@@ -1,9 +1,9 @@
+import { useQueue } from "@/Hooks";
+
 const RfqPage = () => {
-  return (
-    <>
-      <h2>RFQ</h2>
-    </>
-  );
+  const rfqQueue = useQueue("rfq");
+  console.log(rfqQueue);
+  return <h2>RFQ</h2>;
 };
 
 export default RfqPage;
